Add files via upload
Part 1. Develop a UNIX based Shell Script to automate management tasks for a Linux Operating System. The solution will be a menu driven program that provides the systems administrator options to run the following automated tasks.
Part 2. In this assessment you will build a peer to peer network between two Linux machines. A Linux server deployed as a NFS and HTTP Server, a Linux client deployed as a NFS and HTTP Client